Refactor recipes

+++ dnsmanager/recipes.py
@@ -1,4 +1,3 @@
-
 from .models import AddressRecord
 from .models import CanonicalNameRecord
 from .models import MailExchangeRecord
@@ -6,51 +5,84 @@ from .models import NameServerRecord
 from .models import TextRecord
 from .models import ServiceRecord
 from .settings import ZONE_DEFAULTS
-from .signals import zone_fully_saved_signal
 
 
 class Recipe(object):
     """ Custom Zone Recipe """
-
     def __init__(self, zone):
         self.zone = zone
 
 
-class GoogleApps(Recipe):
+class NameServerRecipe(Recipe):
+    """ Superclass for Name Server Recipe """
+    data = None
+
+    def __init__(self, zone):
+        super(NameServerRecipe, self).__init__(zone)
+        self.set_ns(self.data)
+
+    def set_ns(self, data):
+        if data is not None:
+            # Remove existing NS
+            NameServerRecord.objects.filter(zone=self.zone).delete()
+            for d in data:
+                NameServerRecord.objects.get_or_create(zone=self.zone, data=d)
+
+
+class CnameRecipe(Recipe):
+    """ Superclass for CNAME Recipe """
+
+    data = None
+
+    def __init__(self, zone):
+        super(CnameRecipe, self).__init__(zone)
+        self.set_cname(self.data)
+
+    def set_cname(self, data):
+        if data is not None:
+            for d, t in data:
+                CanonicalNameRecord.objects.get_or_create(zone=self.zone, data=d, target=t)
+
+
+class MxRecipe(Recipe):
+    """ Superclass for MX Recipe """
+
+    data = None
+
+    def __init__(self, zone):
+        super(MxRecipe, self).__init__(zone)
+        self.set_mx(self.data)
+
+    def set_mx(self, data):
+        if data is not None:
+            # Remove existing MX
+            MailExchangeRecord.objects.filter(zone=self.zone).delete()
+            for p, d in data:
+                MailExchangeRecord.objects.get_or_create(zone=self.zone, data=d, priority=p)
+
+
+class GoogleApps(CnameRecipe, MxRecipe):
+
+    data_cname = [
+        ('calendar', 'ghs.googlehosted.com.'),
+        ('docs', 'ghs.googlehosted.com.'),
+        ('mail', 'ghs.googlehosted.com.'),
+        ('sites', 'ghs.googlehosted.com.'),
+        ('video', 'ghs.googlehosted.com.'),
+    ]
+
+    data_mx = [
+        ('10', 'aspmx.l.google.com.'),
+        ('20', 'alt1.aspmx.l.google.com.'),
+        ('20', 'alt2.aspmx.l.google.com.'),
+        ('30', 'alt3.aspmx.l.google.com.'),
+        ('30', 'alt4.aspmx.l.google.com.'),
+    ]
 
     def __init__(self, zone):
         super(GoogleApps, self).__init__(zone)
-        self.set_cname()
-        self.set_mx()
-
-    def set_cname(self):
-
-        data = [
-            ('calendar', 'ghs.googlehosted.com.'),
-            ('docs', 'ghs.googlehosted.com.'),
-            ('mail', 'ghs.googlehosted.com.'),
-            ('sites', 'ghs.googlehosted.com.'),
-            ('video', 'ghs.googlehosted.com.'),
-        ]
-
-        for d, t in data:
-            CanonicalNameRecord.objects.get_or_create(zone=self.zone, data=d, target=t)
-
-    def set_mx(self):
-
-        data = [
-            ('10', 'aspmx.l.google.com.'),
-            ('20', 'alt1.aspmx.l.google.com.'),
-            ('20', 'alt2.aspmx.l.google.com.'),
-            ('30', 'alt3.aspmx.l.google.com.'),
-            ('30', 'alt4.aspmx.l.google.com.'),
-        ]
-
-        # Remove existing MX
-        MailExchangeRecord.objects.filter(zone=self.zone).delete()
-
-        for p, d in data:
-            MailExchangeRecord.objects.get_or_create(zone=self.zone, data=d, priority=p)
+        self.set_cname(self.data_cname)
+        self.set_mx(self.data_mx)
 
 
 class RemovePerRecordTtls(Recipe):
@@ -69,7 +101,6 @@ class ResetZoneDefaults(Recipe):
     """ Reset Zone TTLs, expiry etc """
     def __init__(self, zone):
         super(ResetZoneDefaults, self).__init__(zone)
-
         self.zone.refresh = ZONE_DEFAULTS['refresh']
         self.zone.retry = ZONE_DEFAULTS['retry']
         self.zone.expire = ZONE_DEFAULTS['expire']
@@ -81,37 +112,3 @@ class ResetZoneDefaults(Recipe):
 class ReSave(Recipe):
     """ Force resave of the zone """
     pass  # Null recipe
-
-
-class NameServerRecipe(Recipe):
-    """ Superclass for Name Server Recipe """
-    data = None
-
-    def __init__(self, zone):
-        super(NameServerRecipe, self).__init__(zone)
-        self.set_ns()
-
-    def set_ns(self):
-
-        # Remove existing NS
-        NameServerRecord.objects.filter(zone=self.zone).delete()
-
-        for d in self.data:
-            NameServerRecord.objects.get_or_create(zone=self.zone, data=d)
-
-
-class MxRecipe(Recipe):
-
-    data = None
-
-    def __init__(self, zone):
-        super(MxRecipe, self).__init__(zone)
-        self.set_mx()
-
-    def set_mx(self):
-
-        # Remove existing MX
-        MailExchangeRecord.objects.filter(zone=self.zone).delete()
-
-        for p, d in self.data:
-            MailExchangeRecord.objects.get_or_create(zone=self.zone, data=d, priority=p)
